Move test for skipped metadata extraction to `TestExtractDsMeta`
